requirements.txt
python-dotenv
pydantic_settings
fastapi
uvicorn
passlib
cryptography
sqlalchemy
alembic